Move getTypeDeclaration() to AbstractRubyCodegen
